Upated minimal store with native store changes. Optimized a little minimal loader.

+++ src/store/minimal/min_item_vector.h
@@ -38,7 +38,7 @@ public:
   bool isAtomic() const { return false; }
   bool isPul() const    { return false; }
 
-  ulong size() const { return theItems.size(); }
+  ulong size() const { return (ulong)theItems.size(); }
 
   bool empty() const { return theItems.empty(); }
